requirements.txt
asgiref==3.5.2
click==8.1.3
colorama==0.4.5
dj-database-url==1.0.0
Django==4.1.2
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.14.0
gunicorn==20.1.0
joblib==1.2.0
nltk==3.7
numpy==1.23.3
pytz==2022.4
regex==2022.9.13
sqlparse==0.4.3
textblob==0.17.1
torch==1.12.1
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.4
whitenoise==6.2.0
